app/requirements.txt
flask
line-bot-sdk
gunicorn
sqlalchemy
flask-sqlalchemy
flask-migrate
pymysql